{{cookiecutter.project_slug}}/requirements/base.txt
{% if cookiecutter.use_python2 == 'n' -%}
# Wheel 0.25+ needed to install certain packages on CPython 3.5+
# like Pillow and psycopg2
# See http://bitly.com/wheel-building-fails-CPython-35
# Verified bug on Python 3.5.1
wheel==0.29.0
{%- endif %}

# Bleeding edge Django
django==1.9.6

# Configuration
django-environ==0.4.0
django-secure==1.0.1
{% if cookiecutter.use_whitenoise == 'y' -%}
whitenoise==3.0
{%- endif %}


# Forms
django-braces==1.8.1
django-crispy-forms==1.6.0
django-floppyforms==1.6.2
django-betterforms==1.1.4

# Models
django-model-utils==2.5

# Images
Pillow==3.2.0

# For user registration, either via email or social
# Well-built with regular release cycles!
django-allauth==0.25.2

{% if cookiecutter.windows == 'y' -%}
# On Windows, you must download/install psycopg2 manually
# from http://www.lfd.uci.edu/~gohlke/pythonlibs/#psycopg
{% else %}
# Python-PostgreSQL Database Adapter
psycopg2==2.6.1
{%- endif %}

# Unicode slugification
unicode-slugify==0.1.3
django-autoslug==1.9.3

# Time zones support
pytz==2016.4

# Redis support
django-redis==4.4.2
redis>=2.10.0

{% if cookiecutter.use_celery == "y" %}
celery==3.1.23
{% endif %}

# Your custom requirements go here
